feat(web): add UI to upload game files
This allows us to deploy a public static page
that does not require manually adding the `.pk3` files
to the HTTP server's directory.
This way the page can be used by regular people.

Additionally, in the future this might simplify
the development workflow of reviewing MRs,
because this does not require downloading any files
to run the preview build.

The approach is taken from
- https://github.com/jdarpinian/ioq3/blob/7801abd5ac510a889087ad62473c54dbe509e106/code/web/index.html#L614-L660
- https://github.com/WofWca/quake3.xdc/blob/be18795b56a0e408514e78201a731f38c1a92bc6/upload-full-version.html

Co-authored-by: James Darpinian <jdarpinian@gmail.com>